refactor(lexer): move `not_operator_position` into separate module (#26577)
Pure refactor. No code changes, just moving things around.

First part of an effort to split up the massive `disambiguate` module.

`not_operator_position` (previously called `prev_is_regex` before #26576) is a separate self-contained subsystem, so move it, and its helper functions into a separate module.

Also move tests related to `not_operator_position` into that directory, so they sit next to the code they test.

There are 5 tests which cover both `not_operator_position` and other disambiguation decisions in a single test:

- `long_walks_resolve_exactly`
- `escaped_identifiers_are_identifiers_in_every_walk`
- `type_context_before_a_declaration_keyword`
- `gt_run_heads_after_asi_and_this`
- `relational_heads_before_a_balanced_run`

Have left these in the main `disambiguate/tests.rs` file.

I got Claude to do the moving of the tests - too laborious to do by hand. Claude was also responsible for classifying which tests relate to what, and so _which_ tests to move. I have thoroughly checked the first - made sure no tests have been changed in any way, and that none have been lost in the process. I have not checked Claude's logic on the 2nd, but he seems to know what he's doing! This feels like an appropriate level of oversight to me, since much of the disambiguate logic and tests was produced originally with a heavy lean on AI.
